+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [2.0.1] - 2025-01-31
+### Changed
+- IDataService 
+	- Added GetOne
+	- Added tracking option to Get
+	- Added tracking option to GetAsQueryable
+	- Added tracking option to FindAsQueryable
+- Updated packages
+
+## [2.0.0] - 2025-01-17
+### Added
+- IEntity
+- IEntityRepo
+- BaseEntityRepo
+- BaseDataService
+- DataService
+- DataServiceMiddlewareExtension
+### Removed
+- IGenericService
+- AbstractGenericService
+- AbstractDataService
+### Changed
+- IDataService
+- Updated packages
+
 ## [1.0.3] - 2024-09-07
 ### Changed
 - IDataService